requirements.txt
adal==1.2.2
asn1crypto==0.24.0
astroid==2.2.5
azure-common==1.1.23
azure-mgmt-cognitiveservices==5.0.0
azure-mgmt-resource==3.0.0
azure-mgmt-search==2.1.0
azure-mgmt-storage==4.0.0
azure-mgmt-web==0.42.0
azure-storage-blob==2.0.1
azure-storage-common==2.0.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
colorama==0.4.1
cryptography==2.7
Flask==1.1.0
Flask-WTF==0.14.2
idna==2.8
isodate==0.6.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.1
MarkupSafe==1.1.1
mccabe==0.6.1
msrest==0.6.8
msrestazure==0.6.1
oauthlib==3.0.2
pycparser==2.19
PyJWT==1.7.1
pylint==2.3.1
python-dateutil==2.8.0
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
typed-ast==1.4.0
urllib3==1.25.3
Werkzeug==0.15.4
wrapt==1.11.2
WTForms==2.2.1
